Fix #7 change variable name to disable plugin
A breaking change (potentially a bug?) in Neovim has meant that the plugin is no longer recognised for the filetype; although the plugin continues to work in Vim.  The disabling option provided in :help write-filetype-plugin now appears to disable this plugin by default if the help file version is used without editing the buffer variable name.  Changing the variable name prevents this conflict and enables the plugin.